+++ lib/swid/ir/transforms/instantiateAllGenericsAs.freezed.dart
