+++ lib/swid/ir/transforms/instantiateAllGenericsAsDynamic.freezed.dart
