+++ lib/swid/ir/transforms/instantiateGeneric.freezed.dart
